pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "wax_ai"
version = "0.0.1"
description = "WAX AI project for machine learning and artificial intelligence."
authors = [
  { name = "Jakob Nilsson" },
]

readme = "README.md"

classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pip",
    "python-dotenv",
    "ruff",
    "scikit-learn>=1.5.0",
    "seaborn>=0.13.0",
    "joblib>=1.4.0",
    "torch>=2.0.0",
    "tensorflow>=2.15.0",
    "transformers>=4.35.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "streamlit>=1.30.0",
    "alembic>=1.13.0",
]
requires-python = "~=3.13.0"


[tool.ruff]
line-length = 99
src = ["app"]
include = ["pyproject.toml", "app/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["app"]
force-sort-within-sections = true

